[switch to R3] Add ralph3_vip plugin. (#495)
This PR provides a new plugin for VIPs, which is kind of different than the rest, because it is more of a view than a plugin launched from comand line. It basically provides a dedicated endpoint, subscribes it to `refreshVipEvent` in Hermes and processes received data in real time, so in contrast to "normal" plugins, there's no distinction when it starts/finishes.

It also makes `get_unknown_service_env` function more generic, and moves it  to `plugins.collect.utils` (from `ralph3_asset` plugin).